Extract agent-view hotkey wiring into useAgentHotkeys
Moves the four hotkey effects, the cycleAgent helper, the palette open
state, and the paletteActions list out of agents-view.tsx and into
apps/web/src/hooks/use-agent-hotkeys.ts. agents-view now consumes the
behavior through a single hook call and renders <CommandPalette /> with
the returned values — net -90 LOC in agents-view.

No behavior change.
